#72: Reorganize bulletin parsers and serializers for clearer package structure.
