+++ src/main/java/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/TAFBulletinIWXXMParser.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm.v2_1.taf;
+package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
 
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoField;
@@ -10,8 +10,6 @@ import fi.fmi.avi.converter.AviMessageSpecificConverter;
 import fi.fmi.avi.converter.ConversionException;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.iwxxm.bulletin.AbstractBulletinIWXXMParser;
-import fi.fmi.avi.converter.iwxxm.bulletin.BulletinProperties;
-import fi.fmi.avi.converter.iwxxm.bulletin.MeteorologicalBulletinIWXXMScanner;
 import fi.fmi.avi.model.bulletin.BulletinHeading;
 import fi.fmi.avi.model.taf.TAF;
 import fi.fmi.avi.model.taf.TAFBulletin;
