+++ src/main/java/fi/fmi/avi/converter/iwxxm/conf/IWXXMSpaceWeatherConverter.java
@@ -6,9 +6,9 @@ import org.w3c.dom.Document;
 
 import fi.fmi.avi.converter.AviMessageSpecificConverter;
 import fi.fmi.avi.converter.iwxxm.AbstractIWXXMSerializer;
-import fi.fmi.avi.converter.iwxxm.bulletin.BulletinIWXXMDOMSerializer;
-import fi.fmi.avi.converter.iwxxm.bulletin.BulletinIWXXMStringSerializer;
-import fi.fmi.avi.converter.iwxxm.v3_0.swx.SpaceWeatherBulletinIWXXMParser;
+import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.BulletinIWXXMDOMSerializer;
+import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.BulletinIWXXMStringSerializer;
+import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.SpaceWeatherBulletinIWXXMParser;
 import fi.fmi.avi.converter.iwxxm.v3_0.swx.SpaceWeatherIWXXMParser;
 import fi.fmi.avi.converter.iwxxm.v3_0.swx.SpaceWeatherIWXXMSerializer;
 import fi.fmi.avi.model.swx.SpaceWeatherAdvisory;
