+++ src/main/resources/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/collect-template.xml
