+++ src/test/java/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/GenericBulletinParserTest.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm.bulletin;
+package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -21,7 +21,6 @@ import fi.fmi.avi.converter.AviMessageConverter;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.ConversionResult;
 import fi.fmi.avi.converter.iwxxm.IWXXMTestConfiguration;
-import fi.fmi.avi.converter.iwxxm.bulletin.generic.IWXXMGenericBulletinScanner;
 import fi.fmi.avi.converter.iwxxm.conf.IWXXMConverter;
 import fi.fmi.avi.model.bulletin.BulletinHeading;
 import fi.fmi.avi.model.bulletin.GenericMeteorologicalBulletin;
@@ -37,36 +36,36 @@ public class GenericBulletinParserTest {
     private AviMessageConverter converter;
 
     private Document getBulletinDocument(final String filename) throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+        final DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
         dbf.setNamespaceAware(true);
         dbf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
-        DocumentBuilder db = dbf.newDocumentBuilder();
+        final DocumentBuilder db = dbf.newDocumentBuilder();
         return db.parse(GenericBulletinParserTest.class.getResourceAsStream(filename));
     }
 
     @Test
     public void testScanner() throws Exception {
-        BulletinProperties properties = new BulletinProperties();
-        IWXXMGenericBulletinScanner scanner = new IWXXMGenericBulletinScanner();
+        final BulletinProperties properties = new BulletinProperties();
+        final IWXXMGenericBulletinScanner scanner = new IWXXMGenericBulletinScanner();
         scanner.collectBulletinProperties(this.getBulletinDocument("taf-bulletin.xml"), properties, ConversionHints.EMPTY);
         assertTrue(properties.contains(BulletinProperties.Name.HEADING));
-        Optional<BulletinHeading> heading = properties.get(BulletinProperties.Name.HEADING, BulletinHeading.class);
+        final Optional<BulletinHeading> heading = properties.get(BulletinProperties.Name.HEADING, BulletinHeading.class);
 
         assertTrue(properties.contains(BulletinProperties.Name.MESSAGE));
     }
 
     @Test
     public void testParserWithTAF() throws Exception {
-        Document input = this.getBulletinDocument("taf-bulletin.xml");
-        ConversionResult<GenericMeteorologicalBulletin> result = this.converter.convertMessage(input, IWXXMConverter.IWXXM21_DOM_TO_GENERIC_BULLETIN_POJO,
+        final Document input = this.getBulletinDocument("taf-bulletin.xml");
+        final ConversionResult<GenericMeteorologicalBulletin> result = this.converter.convertMessage(input, IWXXMConverter.IWXXM21_DOM_TO_GENERIC_BULLETIN_POJO,
                 ConversionHints.EMPTY);
         assertEquals(ConversionResult.Status.SUCCESS, result.getStatus());
     }
 
     @Test
     public void testParserWithSIGMET() throws Exception {
-        Document input = this.getBulletinDocument("sigmet-bulletin.xml");
-        ConversionResult<GenericMeteorologicalBulletin> result = this.converter.convertMessage(input, IWXXMConverter.IWXXM21_DOM_TO_GENERIC_BULLETIN_POJO,
+        final Document input = this.getBulletinDocument("sigmet-bulletin.xml");
+        final ConversionResult<GenericMeteorologicalBulletin> result = this.converter.convertMessage(input, IWXXMConverter.IWXXM21_DOM_TO_GENERIC_BULLETIN_POJO,
                 ConversionHints.EMPTY);
         assertEquals(ConversionResult.Status.SUCCESS, result.getStatus());
     }
