+++ src/test/java/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/SpaceWeatherBulletinParserTest.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm.v2_1;
+package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -21,54 +21,54 @@ import fi.fmi.avi.converter.AviMessageConverter;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.ConversionResult;
 import fi.fmi.avi.converter.iwxxm.IWXXMTestConfiguration;
-import fi.fmi.avi.converter.iwxxm.bulletin.BulletinProperties;
-import fi.fmi.avi.converter.iwxxm.bulletin.MeteorologicalBulletinIWXXMScanner;
 import fi.fmi.avi.converter.iwxxm.conf.IWXXMConverter;
 import fi.fmi.avi.model.bulletin.BulletinHeading;
-import fi.fmi.avi.model.taf.TAF;
-import fi.fmi.avi.model.taf.TAFBulletin;
+import fi.fmi.avi.model.swx.SpaceWeatherAdvisory;
+import fi.fmi.avi.model.swx.SpaceWeatherBulletin;
+import fi.fmi.avi.model.swx.SpaceWeatherPhenomenon;
 
 /**
  * Created by rinne on 19/07/17.
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = IWXXMTestConfiguration.class, loader = AnnotationConfigContextLoader.class)
-public class TAFBulletinParserTest {
+public class SpaceWeatherBulletinParserTest {
 
     @Autowired
     private AviMessageConverter converter;
 
     private Document getBulletinDocument(final String filename) throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+        final DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
         dbf.setNamespaceAware(true);
         dbf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
-        DocumentBuilder db = dbf.newDocumentBuilder();
-        return db.parse(TAFBulletinParserTest.class.getResourceAsStream(filename));
+        final DocumentBuilder db = dbf.newDocumentBuilder();
+        return db.parse(SpaceWeatherBulletinParserTest.class.getResourceAsStream(filename));
     }
 
     @Test
     public void testScanner() throws Exception {
-        assertTrue(this.converter.isSpecificationSupported(IWXXMConverter.IWXXM21_DOM_TO_TAF_POJO));
-        BulletinProperties properties = new BulletinProperties();
-        MeteorologicalBulletinIWXXMScanner<TAF, TAFBulletin> scanner = new MeteorologicalBulletinIWXXMScanner<>();
-        scanner.setMessageConverter(converter.getConverter(IWXXMConverter.IWXXM21_DOM_TO_TAF_POJO));
-        scanner.collectBulletinProperties(this.getBulletinDocument("taf-bulletin.xml"), properties, ConversionHints.EMPTY);
+        assertTrue(this.converter.isSpecificationSupported(IWXXMConverter.IWXXM30_DOM_TO_SPACE_WEATHER_POJO));
+        final BulletinProperties properties = new BulletinProperties();
+        final MeteorologicalBulletinIWXXMScanner<SpaceWeatherAdvisory, SpaceWeatherBulletin> scanner = new MeteorologicalBulletinIWXXMScanner<>();
+        scanner.setMessageConverter(converter.getConverter(IWXXMConverter.IWXXM30_DOM_TO_SPACE_WEATHER_POJO));
+        scanner.collectBulletinProperties(this.getBulletinDocument("swx-bulletin.xml"), properties, ConversionHints.EMPTY);
         assertTrue(properties.contains(BulletinProperties.Name.HEADING));
-        Optional<BulletinHeading> heading = properties.get(BulletinProperties.Name.HEADING, BulletinHeading.class);
+        final Optional<BulletinHeading> heading = properties.get(BulletinProperties.Name.HEADING, BulletinHeading.class);
         assertEquals(31, heading.get().getBulletinNumber());
         assertTrue(properties.contains(BulletinProperties.Name.MESSAGE));
     }
 
     @Test
     public void testParser() throws Exception {
-        Document input = this.getBulletinDocument("taf-bulletin.xml");
-        ConversionResult<TAFBulletin> result = this.converter.convertMessage(input, IWXXMConverter.WMO_COLLECT_DOM_TO_TAF_BULLETIN_POJO, ConversionHints.EMPTY);
+        final Document input = this.getBulletinDocument("swx-bulletin.xml");
+        final ConversionResult<SpaceWeatherBulletin> result = this.converter.convertMessage(input, IWXXMConverter.WMO_COLLECT_DOM_TO_SWX_BULLETIN_POJO,
+                ConversionHints.EMPTY);
         assertEquals(ConversionResult.Status.SUCCESS, result.getStatus());
         if (result.getConvertedMessage().isPresent()) {
-            TAFBulletin bulletin = result.getConvertedMessage().get();
-            assertEquals(2, bulletin.getMessages().size());
-            TAF mesg = bulletin.getMessages().get(0);
-            assertEquals(26.0, mesg.getBaseForecast().get().getSurfaceWind().get().getWindGust().get().getValue(), 0.0001);
+            final SpaceWeatherBulletin bulletin = result.getConvertedMessage().get();
+            assertEquals(1, bulletin.getMessages().size());
+            final SpaceWeatherAdvisory mesg = bulletin.getMessages().get(0);
+            assertEquals(SpaceWeatherPhenomenon.fromWMOCodeListValue("http://codes.wmo.int/49-2/SpaceWxPhenomena/HF_COM_MOD"), mesg.getPhenomena().get(0));
         }
     }
