+++ src/test/resources/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/swx-bulletin.xml
